documentation for timing & keys modules
