+++ Cargo.toml
@@ -1,7 +1,9 @@
 [package]
-name = "voyager"
+name = "voyager_client"
 version = "0.1.0"
 edition = "2021"
+description = "Client for interacting with NASA's Open APIs"
+license = "MIT"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
